Crate tls_initializer

Logic for generating Thread-Local Storage (TLS) data image for TLS areas.

The two key types are:

TlsInitializer: a “factory” that maintains a list of loaded TLS sections in order to correctly generate new TLS data images.

TlsDataImage: a generated TLS data image that can be used as the TLS area for a single task.

TODO FIXME: currently we are unsure of the virt_addr_values calculated for TLS sections on aarch64. The placement of those sections in the TLS data image is correct, but relocations against them may not be.

Structs

TlsDataImage

An initialized TLS area data image ready to be used by a new task.

TlsInitializer

A “factory” that creates Thread-Local Storage (TLS) data images, which are used to initialize a new Task’s TLS area.

Enums

TlsInitializerError

Errors that may occur when adding TLS sections to a TlsInitializer.